backend/requirements.txt
fastapi
uvicorn[standard]
pyautogen==0.2.9      # multi-agent runtime, pinned version
openai
qdrant-client
redis
minio
python-multipart      # file uploads
tiktoken
python-dotenv
pytest
pytest-asyncio        # для async тестов
websockets            # для WebSocket тестов
pydantic>=2           # требуется версия 2+
pydantic-settings     # для BaseSettings в Pydantic v2
pdfminer.six         # извлечение текста из PDF
python-docx          # docx to text (если нет – ok, fallback)
tqdm                  # прогресс-бар для индексатора
prometheus-fastapi-instrumentator  # Prometheus metrics
grpcio
grpcio-tools
grpclib
protobuf>=5.26.0      # Обновлённая версия protobuf
opentelemetry-api
opentelemetry-sdk
sse-starlette                       # Server-Sent Events
opentelemetry-exporter-otlp
structlog
